Added the most arbitrary feature ever
You'll never believe this, but the production company came to me and demanded that the "core gameplay should still involve controlling a rolling ball". I have therefore added in the ability to push the ball left and right after throwing it, meaning the core gameplay now involves controlling a rolling ball. Does this add anything meaningful to the game? Not really! Is it in the spirit of iterating on an idea to find better ideas that lead to more fun? Also not really! But what the boss says goes, even if the boss never actually puts it in writing once before now deciding that it's a must-have. Hopefully this resolves all of the "can't control a rolling ball" concerns that many players were surely having.